Add ClientApplication subdir to break up AppConfig subdir
